+++ lib/app/app.dart
@@ -1,14 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:health_diary/modules/debug/screens/settings/debug_settings_screen.dart';
+import 'package:health_diary/services/theme_service.dart';
 
 class App extends StatelessWidget {
-  const App({Key? key}) : super(key: key);
+  final ThemeService themeService;
+
+  const App({required this.themeService, Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      theme: ThemeData(primarySwatch: Colors.blue),
+      theme: themeService.lightTheme,
+      darkTheme: themeService.dartTheme,
       home: const DebugSettingsScreen(),
     );
   }