+++ test/modules/debug/screens/settings/debug_settings_screen_test.dart
@@ -1,3 +1,4 @@
+import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
 import 'package:health_diary/modules/debug/screens/settings/debug_settings_screen.dart';
@@ -22,6 +23,13 @@ void main() {
     testGoldens('golden test', (tester) async {
       await tester.pumpWidgetBuilder(
         DebugSettingsScreen(wmFactory: (_) => DebugSettingsScreenWM(model)),
+        wrapper: materialAppWrapper(
+          theme: FlexThemeData.light(
+            scheme: FlexScheme.deepBlue,
+            // Use very subtly themed app bar elevation in light mode.
+            appBarElevation: 0.5,
+          ),
+        ),
       );
 
       await multiScreenGolden(tester, 'debug_settings_screen');