+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * build: introduce golden tests
 * build: collect tests coverage
 * feat: introduce exercise widget
+* feat: setup material theme